ThirdPartyResource config store backend (#852)
* ThirdPartyResource config store backend

* missing files

* refactoring done

* fix watcher problem

* fix linter issues

* linter

* move to shared test

* merging #853

* update coverage
