+++ adapter/config/tpr/BUILD
@@ -7,36 +7,31 @@ go_library(
         "config.go",
         "controller.go",
         "conversion.go",
-        "ingressstatus.go",
-        "queue.go",
     ],
     visibility = ["//visibility:public"],
     deps = [
         "//model:go_default_library",
+        "//platform/kube:go_default_library",
         "@com_github_golang_glog//:go_default_library",
+        "@com_github_golang_protobuf//proto:go_default_library",
         "@com_github_hashicorp_go_multierror//:go_default_library",
         "@io_istio_api//:go_default_library",
+        "@io_k8s_apimachinery//pkg/api/errors:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime:go_default_library",
         "@io_k8s_apimachinery//pkg/runtime/schema:go_default_library",
+        "@io_k8s_apimachinery//pkg/runtime/serializer:go_default_library",
+        "@io_k8s_apimachinery//pkg/util/intstr:go_default_library",
         "@io_k8s_apimachinery//pkg/watch:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//pkg/api/v1:go_default_library",
+        "@io_k8s_client_go//pkg/api:go_default_library",
         "@io_k8s_client_go//pkg/apis/extensions/v1beta1:go_default_library",
         "@io_k8s_client_go//plugin/pkg/client/auth/gcp:go_default_library",
         "@io_k8s_client_go//plugin/pkg/client/auth/oidc:go_default_library",
+        "@io_k8s_client_go//rest:go_default_library",
         "@io_k8s_client_go//tools/cache:go_default_library",
         "@io_k8s_client_go//tools/clientcmd:go_default_library",
-        "@io_k8s_client_go//util/flowcontrol:go_default_library",
-        "@io_k8s_ingress//core/pkg/ingress/status:go_default_library",
-        "@io_k8s_ingress//core/pkg/ingress/store:go_default_library",
     ],
-)
-
-filegroup(
-    name = "kubeconfig",
-    srcs = ["config"],
-    visibility = ["//visibility:public"],
 )
 
 go_test(
@@ -46,22 +41,20 @@ go_test(
         "client_test.go",
         "controller_test.go",
         "conversion_test.go",
-        "ingressstatus_test.go",
-        "queue_test.go",
     ],
-    data = [":kubeconfig"] + glob(["testdata/*"]),
+    data = ["//platform/kube:kubeconfig"],
     library = ":go_default_library",
     deps = [
         "//model:go_default_library",
+        "//platform/kube:go_default_library",
         "//proxy:go_default_library",
+        "//test/mock:go_default_library",
         "//test/util:go_default_library",
         "@com_github_golang_glog//:go_default_library",
         "@io_istio_api//:go_default_library",
         "@io_k8s_apimachinery//pkg/apis/meta/v1:go_default_library",
+        "@io_k8s_apimachinery//pkg/util/intstr:go_default_library",
         "@io_k8s_client_go//kubernetes:go_default_library",
-        "@io_k8s_client_go//kubernetes/fake:go_default_library",
-        "@io_k8s_client_go//pkg/api/v1:go_default_library",
         "@io_k8s_client_go//pkg/apis/extensions/v1beta1:go_default_library",
-        "@io_k8s_ingress//core/pkg/ingress/annotations/class:go_default_library",
     ],
 )
