+++ bin/codecov.sh
@@ -6,7 +6,7 @@ echo "" > coverage.txt
 for d in $(go list ./... | grep -v vendor); do
     options="-coverprofile=profile.out -covermode=atomic"
     case $d in
-        istio.io/pilot/platform/kube)
+        istio.io/pilot/adapter/config/tpr)
             echo "Skipping race detection on $d (see https://github.com/istio/pilot/issues/173)"
             ;;
         *)
