+++ cmd/istioctl/BUILD
@@ -10,12 +10,12 @@ go_library(
     ],
     visibility = ["//visibility:private"],
     deps = [
+        "//adapter/config/tpr:go_default_library",
         "//apiserver:go_default_library",
         "//client/proxy:go_default_library",
         "//cmd:go_default_library",
         "//cmd/version:go_default_library",
         "//model:go_default_library",
-        "//platform/kube:go_default_library",
         "//platform/kube/inject:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_golang_glog//:go_default_library",
@@ -47,11 +47,10 @@ go_test(
     ],
     library = ":go_default_library",
     deps = [
+        "//adapter/config/tpr:go_default_library",
         "//apiserver:go_default_library",
         "//client/proxy:go_default_library",
         "//cmd/version:go_default_library",
-        "//model:go_default_library",
-        "//platform/kube:go_default_library",
         "//test/util:go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_googleapis_googleapis//:go_default_library",
