+++ cmd/istioctl/main.go
@@ -31,12 +31,12 @@ import (
 	kubeyaml "k8s.io/apimachinery/pkg/util/yaml"
 	"k8s.io/client-go/pkg/api"
 
+	"istio.io/pilot/adapter/config/tpr"
 	"istio.io/pilot/apiserver"
 	"istio.io/pilot/client/proxy"
 	"istio.io/pilot/cmd"
 	"istio.io/pilot/cmd/version"
 	"istio.io/pilot/model"
-	"istio.io/pilot/platform/kube"
 )
 
 const (
@@ -49,7 +49,7 @@ const (
 type k8sRESTRequester struct {
 	namespace string
 	service   string
-	client    *kube.Client
+	client    *tpr.Client
 }
 
 // Request wraps Kubernetes specific requester to provide the proper
@@ -58,7 +58,7 @@ func (rr *k8sRESTRequester) Request(method, path string, inBody []byte) (int, []
 	return rr.client.Request(rr.namespace, rr.service, method, path, inBody)
 }
 
-func kubeClientFromConfig(kubeconfig string) (*kube.Client, error) {
+func kubeClientFromConfig(kubeconfig string) (*tpr.Client, error) {
 	if kubeconfig == "" {
 		if v := os.Getenv("KUBECONFIG"); v != "" {
 			glog.V(2).Infof("Setting configuration from KUBECONFIG environment variable")
@@ -66,11 +66,11 @@ func kubeClientFromConfig(kubeconfig string) (*kube.Client, error) {
 		}
 	}
 
-	c, err := kube.NewClient(kubeconfig, model.IstioConfigTypes, istioNamespace)
+	c, err := tpr.NewClient(kubeconfig, model.IstioConfigTypes, istioNamespace)
 	if err != nil && kubeconfig == "" {
 		// If no configuration was specified, and the platform
 		// client failed, try again using ~/.kube/config
-		c, err = kube.NewClient(os.Getenv("HOME")+"/.kube/config", model.IstioConfigTypes, istioNamespace)
+		c, err = tpr.NewClient(os.Getenv("HOME")+"/.kube/config", model.IstioConfigTypes, istioNamespace)
 	}
 	if err != nil {
 		return nil, multierror.Prefix(err, "failed to connect to Kubernetes API.")
@@ -138,7 +138,7 @@ istioctl mixer command documentation.
 				apiClient = proxy.NewConfigClient(&proxy.BasicHTTPRequester{
 					BaseURL: istioConfigAPIService,
 					Client:  &http.Client{Timeout: 60 * time.Second},
-					Version: kube.IstioResourceVersion,
+					Version: tpr.IstioResourceVersion,
 				})
 			}
