+++ cmd/pilot/BUILD
@@ -5,6 +5,7 @@ go_library(
     srcs = ["main.go"],
     visibility = ["//visibility:private"],
     deps = [
+        "//adapter/config/tpr:go_default_library",
         "//apiserver:go_default_library",
         "//cmd:go_default_library",
         "//cmd/version:go_default_library",
