+++ platform/kube/conversion_test.go
@@ -26,12 +26,6 @@ import (
 var (
 	domainSuffix = "company.com"
 
-	camelKabobs = []struct{ in, out string }{
-		{"ExampleNameX", "example-name-x"},
-		{"example1", "example1"},
-		{"exampleXY", "example-x-y"},
-	}
-
 	protocols = []struct {
 		name  string
 		proto v1.Protocol
@@ -51,69 +45,11 @@ var (
 	}
 )
 
-func TestCamelKabob(t *testing.T) {
-	for _, tt := range camelKabobs {
-		s := camelCaseToKabobCase(tt.in)
-		if s != tt.out {
-			t.Errorf("camelCaseToKabobCase(%q) => %q, want %q", tt.in, s, tt.out)
-		}
-	}
-}
-
 func TestConvertProtocol(t *testing.T) {
 	for _, tt := range protocols {
 		out := convertProtocol(tt.name, tt.proto)
 		if out != tt.out {
 			t.Errorf("convertProtocol(%q, %q) => %q, want %q", tt.name, tt.proto, out, tt.out)
-		}
-	}
-}
-
-func TestDecodeIngressRuleName(t *testing.T) {
-	cases := []struct {
-		ingressName      string
-		ingressNamespace string
-		ruleNum          int
-		pathNum          int
-	}{
-		{"myingress", "test", 0, 0},
-		{"myingress", "default", 1, 2},
-		{"my-ingress", "test-namespace", 1, 2},
-		{"my-cool-ingress", "new-space", 1, 2},
-	}
-
-	for _, c := range cases {
-		encoded := encodeIngressRuleName(c.ingressName, c.ingressNamespace, c.ruleNum, c.pathNum)
-		ingressName, ingressNamespace, ruleNum, pathNum, err := decodeIngressRuleName(encoded)
-		if err != nil {
-			t.Errorf("decodeIngressRuleName(%q) => error %v", encoded, err)
-		}
-		if ingressName != c.ingressName || ingressNamespace != c.ingressNamespace ||
-			ruleNum != c.ruleNum || pathNum != c.pathNum {
-			t.Errorf("decodeIngressRuleName(%q) => (%q, %q, %d, %d), want (%q, %q, %d, %d)",
-				encoded,
-				ingressName, ingressNamespace, ruleNum, pathNum,
-				c.ingressName, c.ingressNamespace, c.ruleNum, c.pathNum,
-			)
-		}
-	}
-}
-
-func TestIsRegularExpression(t *testing.T) {
-	cases := []struct {
-		s       string
-		isRegex bool
-	}{
-		{"/api/v1/", false},
-		{"/api/v1/.*", true},
-		{"/api/.*/resource", true},
-		{"/api/v[1-9]/resource", true},
-		{"/api/.*/.*", true},
-	}
-
-	for _, c := range cases {
-		if isRegularExpression(c.s) != c.isRegex {
-			t.Errorf("isRegularExpression(%q) => %v, want %v", c.s, !c.isRegex, c.isRegex)
 		}
 	}
 }
