+++ test/integration/infra.go
@@ -28,9 +28,9 @@ import (
 	meta_v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 
 	proxyconfig "istio.io/api/proxy/v1/config"
+	"istio.io/pilot/adapter/config/tpr"
 	"istio.io/pilot/cmd"
 	"istio.io/pilot/model"
-	"istio.io/pilot/platform/kube"
 	"istio.io/pilot/platform/kube/inject"
 	"istio.io/pilot/test/util"
 )
@@ -270,7 +270,7 @@ func (infra *infra) applyConfig(inFile string, data map[string]string, typ strin
 		return err
 	}
 
-	istioClient, err := kube.NewClient(kubeconfig, model.IstioConfigTypes, infra.Namespace)
+	istioClient, err := tpr.NewClient(kubeconfig, model.IstioConfigTypes, infra.Namespace)
 	if err != nil {
 		return err
 	}
